+++ src/main/java/attendance/domain/analysis/AttendanceRecord.java
@@ -1,8 +1,7 @@
-package attendance;
+package attendance.domain.analysis;
 
-import attendance.domain.*;
-
-import java.time.LocalDate;
+import attendance.domain.model.*;
+import attendance.domain.enums.*;
 import java.util.List;
 import java.util.stream.Collectors;
 import java.util.Comparator;
