+++ src/main/java/attendance/domain/enums/AttendanceStatus.java
@@ -1,4 +1,4 @@
-package attendance.domain;
+package attendance.domain.enums;
 
 public enum AttendanceStatus {
     PRESENT("출석"),
